l2geth: make block hashes deterministic
The blockhashes can be non deterministic because there is not a global
consensus in the existing deployment of the optimism network, instead
each node runs with `--dev 0` meaning that it auto mines transactions
itself. Previously, block hashes were non deterministic due to a
different clique signing key being used but that was fixed so that
a deterministic key is used for all nodes.

This fixes the possibility of the block extradata being different which
would result in different block hashes. The extradata is hard coded
to be the same value as the release @eth-optimism/l2geth@0.4.6. In
version 0.4.7, the underlying Dockerfile pulled in a patch release of
the Go runtime which caused the extradata field to become different.
The extradata field by default is the version of the software (which
cannot change until the next regenesis), "geth", the go runtime version
and the operating system.

This will require a resync to make block hashes deterministic across the
network.