deps: bump to later ethers version (eip1559)
Using EIP-1559 should reduce transaction fees as using
legacy transactions are more expensive based on the
EIP-1559 mechanism. This bumps the version of ethers
across all of the packages.

This also regenerates the `yarn.lock` which upgrades
deps across the codebase. It was not building until
this was done due to mismatched versions in the
`yarn.lock`