batch-submitter: handle error case explicitly
It was possible to return from a function directly
into destructuring an array when the return value
was polymorphic in either being an array or undefined.
The undefined case is handled explicitly now and an
error is thrown with a useful error message instead of
something that would require looking at the source code
and knowing nodejs error types to find out what the
problem was.