l2geth: fix for `gas-oracle` race condition
Prevent the error being returned in consensus for when
the user does not have enough balance. The policy level
check *must* filter out transactions that do not have
enough gas to pay for their execution. The policy
level check happens in the `SyncService`. This change
gracefully handles the race condition where a user transaction
passes the policy level check as a `gas-oracle` transaction
is executing and updating the gas prices. If the gas price
goes up and the user transaction no longer can afford execution,
then it will return an error from consensus.