fix: don't round down average gas per epoch
Also corrects the targetGps constant in the test harness. The constant
was chosen to be the expected Gps of the initial burst, causing the Gps
to stablize in the second part of the test at the floor.

Adding a check that the average measured gas price matches our desired
constant prevents this from silently passing.