feat: surface miner nonce too high errors
This commit exposes the ErrNonceTooHigh at the execution level to
external users. Previously the error was mapped to ErrNonceTooLow to
match the mempool behavior. As this is a new error, and one not
currently exposed in L1, this could be a breaking change for some
applications depending on how the error messages are parsed.