fix(batch-submitter): clear state root batches
Fixes a bug where both the sequencer and proposer main loops were
attempting to clear the pending transactions from the batch-tx
submitter's wallet. The impact is that we may have been overspending on
fees due to conflicting/reverting state batches. However the impact
overall should be minor given the relative size of state batches in
comparison to tx-batches.