l2geth: berlin hardfork precompile implementations
Add in the berlin precompile maps. This code is adopted from
upstream.

This includes the implementation of EIP-2565 which
makes the modular exponentiation precompile found
at `address(5)` cheaper.

https://eips.ethereum.org/EIPS/eip-2565

The berlin precompile set includes setting the `eip2565`
field to true on the `bigExpMod` struct.

This implementation is directly taken from upstream.
https://github.com/ethereum/go-ethereum/commit/eb21c652c0a9d8f651efc0251cc5797a3328d863