op-chain-ops: add cross domain utils (#3402)
* op-chain-ops: implement pending withdrawal fetching

We need a list of pending withdrawals to be able to
migrate the withdrawal hashes to the new withdrawal
hash scheme. This commit implements the script for
fetching pending withdrawals as well as a unit test
for the function responsible for fetching pending withdrawals.

From within `op-chain-ops`, the script can be ran as follows:

```golang
$ go run cmd/withdrawals/main.go
```

lint

* op-chain-ops: go mod tidy

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Matthew Slipper <me@matthewslipper.com>