op-chain-ops: implement withdrawal hashing (#3469)
* op-chain-ops: implement withdrawal hashing

Implement both the new hashing scheme and the legacy
hashing for withdrawals so that the withdrawals can
be migrated.

Also implement encoding and decoding of the withdrawals

* more tests

* op-chain-ops: code review fixes

* op-chain-ops: fix message passer

Now that the old message passer is being kept in the L2
state, be sure to the use legacy address. The new message
passer is at a different address.