ci-builder: update geth version
Updates the version of `abigen` that is used in the
`ci-builder` so that it produces bindings that do not
parse JSON every time that an instance of the binding
is created.

The change in geth is here: https://github.com/ethereum/go-ethereum/pull/25574

Fixes the additional diff found here: https://github.com/ethereum-optimism/optimism/pull/3680/files

Right now we do not have a guarantee that local devs use
a specific version of `abigen`. After this is merged,
all devs would need to use `v1.10.25`. Devs could use higher
versions but risk running into problems if there are more
changes to `abigen` in the future.