op-chain-ops: add helper function for deployments
Adds a helper function to read the correct deployment
artifacts from disk given a `hardhat.Hardhat` instance.
This code was being repeated throughout the codebase
so its better to consolidate it into a single location.
Now the `DeployConfig` has the L1 contract addresses
required to build a L2 genesis. These addresses do not
need to be present in the JSON file itself and instead
can be read from disk dynamically.

Also add a check function that will error on a config
that is not sane. This can be made more strict in the
future, currently do not want to break anything in the
process of introducing this.