contracts-bedrock: wait for final batch hh task
Implement a simple hardhat task that can be used to
wait for the final batches to be submitted.

Technically, the state batches can only be appended
after the tx batches have been submitted, so we only
need to check for the total elements on the `StateCommitmentChain`
but this checks both the `StateCommitmentChain` and the
`CanonicalTransactionChain`.