go: Remove individual versions for bedrock code
The following packages had their go.mod file removed:
* op-e2e
* op-batcher
* op-bindings
* op-chain-ops
* op-node
* op-proposer
* op-service
* op-signer

They all had cross dependencies previously and used a go.work file
to assume they could be built on the latest code rather than on
the versioned code.

This means that all of these packages are versioned with the root
go.mod. For packages that still have a mod file, they are versioned
independently.