op-chain-ops: more strict checks for config validity
Adds in additional sanity checks for the config validity
to prevent accidental bad configurations of the L2 predeploys.
Ensures that specific values are set in the immutables. Note
this this makes the immutables config specific to optimism,
previously it was implemented in a generic way using a map
so that it could be used with other projects that are doing
similar things with setting contracts in state with their immutables
preset. We could turn it into a struct instead of using a map
which could make it easier to use.

Tests were updated as the data in the tests failed the validation.