op-node: Reproduce invalid advance in unit tests
This reproduces a bug found by mdehoog. This bug causes the system
to halt under certain circustances. Specifically, when generating
batches at the end of the sequence window the batch queue would attempt
to generate empy batches from the next L2 timestamp to the timestamp of
the next L1 origin. If the timestamp of the next L1 origin was less than
the timestamp of the next L2 block time, no batches would be generated
and the internal L1 blocks would be advanced. This would cause L1 origins
to be skipped. This got caught by a consistency check we have in the batch
queue, but loops in an unrecoverable error until this bug is fixed.