migration: update migrated withdrawal gas limit
More accurately computes the gas limit for a migrated
withdrawal. This is useful to prevent the gaslimit
for large transactions from going over the consensus
level block gas limit. It is unlikely for a transaction
to go over the block gas limit, but this is a preventative
fix just in case.