contracts-bedrock: fix standard bridge token pair checking
This ensures that the legacy interface is only used for
legacy tokens and that the modern interface can be used
on the OptimismMintableERC20 token.

Note that this is not the gas efficient way to implement this,
but its ok for now as we are focused on minimal diff changes.

In the future we could refactor this code to ensure that
the ERC165Checker calls a minimal amount of times.