op-chain-ops: Do not manually copy balance in WipePredeployStorage
`db.CreateAccount` copies over the balance, but wipes nonces. The previous version of this code manually copied the balance after calling `db.CreateAccount`, which effectively doubled the predeploy's balance. None of the L2 predeploys carried balances on Goerli, however to prevent ETH from being created and destroyed during the migration this PR removes the manual balance migration and adds a test to ensure that nonces are copied and state is properly wiped.

Fixes CLI-3443