contracts-bedrock: ResourceMetering more readable
Makes the arithmetic in the `ResourceMetering` contract
more readable by using explicit argument syntax when calling
the `Arithmetic` library functions. This syntax is not used
throughout the rest of the codebase, although it should be.
It was determined that it would be too large of a change
to use this syntax throughout the codebase now, but it is
something that we want to do in the future. This adds the
syntax to a few key places in `ResourceMetering` which makes
it much easier to read and will therefore increase auditability
and reviewability of this code.