specs: Update derivation spec to match the current implementation
Dropping frames due to timed out channels or duplication is only done for channels that have not yet been pruned from the channel bank.
Timed out channels are only pruned from the channel-bank if they are prior to a readable channel.
Explicitly state that channels are removed from the channel-bank when they are read.