op-stack: Validate prefixed Environment Variables
This looks for all environment variables with a prefix like OP_NODE
and then validates them against the list of environment variables
provided by the flags for the component. It is not a critical error
to provide an unused environment error, but we do log at a warn level
to suggest to users that the value is not taking an effect.