devnet: unify tooling for how to bring up devnet
The bash script is removed in favor of unifying on the python
code. In the future we should only have 1 way to bring up the devnet
because it is difficult to maintain 2 different ways. It adds
additional CI runs plus the need to add functionality to 2 different
things when contracts are added to the system.