ops: bring back scripts
Bring back `ci-builder` to the monorepo as it should have never
been deleted. Also bring back various tagging scripts. It is not
clear which tagging script is used at the moment, need clarity
on which should be actually brought back before merging this.
`ci-versions.js` is required for changesets, so bring that back
for sure. `update-op-geth.py` is used but unclear about the 2 other
shell scripts.