packages: regenerate yarn.lock
Looks like one of the packages is using an older
version of `core-utils` and is not reflected in
the `yarn.lock`. This commit updates the `yarn.lock`
using the `yarn` command. It also deletes the typescript
types for `bufio` which were used in the legacy batch
submitter but no longer used.