op-chain-ops: cleanup default gas limit
Use a single default value for when configuring the genesis
value of the block gas limit for both L1 and L2 when it is not
set in the deploy config. The value defaults to 30 million when
it is not explicitly set.

This PR is slowly moving changes out of https://github.com/ethereum-optimism/optimism/pull/6123
into their own smaller PRs so that they can be merged in independently
of the final functionality.