op-chain-ops: tx-builder safe types
Implement the types for working with the Safe
tx-builder application. These types should be able
to be round trip serialized into JSON. These types
will be used to create Safe tx-builder bundles
that can be dropped into the UI or can be combined
with additional tooling to send the transactions
to the Safe backend directly or directly interact
with the safe contracts.