contracts-bedrock: upgrade to multichain script
Implements a `forge script` to upgrade to the new L1 contracts.
This assumes that the proxy admin owner is a gnosis safe.

The test can be ran with the following command:

```bash
forge test -vv --contracts scripts/upgrades --mc Multichain --rpc-url $ETH_RPC_URL
```

Where `ETH_RPC_URL` is set to a goerli endpoint. Add `-vvvvv` instead of
`-vv` to see the execution that happens.