Fix dangling modified file on make devnet-up
When starting from a fresh clone of the repo and running

```
make
make cannon-prestate
make devnet-up
```

We are left with a file changed in the tree at

```
packages/contracts-bedrock/deploy-config/devnetL1.json
```

This file is a bit annoying as it should not be checked in but shows as
an outstanding change.  The python script reads the original file, backs
it up, then rewrites that file with a new timestamp.

I'm not entirely certain why the timestamp needs to be rewritten, so
I've left that modification in place.  Instead, this change takes the
approach of moving the original file to a 'template' path, and ignoring
the modified file.  In this way there's no longer any need to constantly
create a backup and have the changes show up in a git diff.